Add interceptor

+++ src/products/products.controller.ts
@@ -7,6 +7,7 @@ import {
   Body,
   ParseUUIDPipe,
   Put,
+  NotFoundException,
 } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
@@ -25,11 +26,19 @@ export class ProductsController {
 
   @Get('/:id')
   getById(@Param('id', new ParseUUIDPipe()) id: string) {
-    return this.productsService.getById(id);
+    const prod = this.productsService.getById(id);
+    if (!prod) {
+      throw new NotFoundException('Product not found');
+    }
+    return prod;
   }
 
   @Delete('/:id')
   deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
+
     this.productsService.delete(id);
     return { success: true };
   }
@@ -44,6 +53,10 @@ export class ProductsController {
     @Param('id', new ParseUUIDPipe()) id: string,
     @Body() productData: UpdateProductDTO,
   ) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
+
     this.productsService.updateById(id, productData);
     return { success: true };
   }
